pyproject.toml
[tool.poetry]
name = "interweb"
version = "0.1.0"
description = ""
authors = ["ilimk <ilimkanu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
django-environ = "^0.10.0"
psycopg2 = "^2.9.6"
black = "^23.7.0"
djangorestframework = "^3.14.0"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
